Divide the test_pr24 test into several test files. Rerun the notebooks with the new version
