+++ tests/test_custom_pulse.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import re
 
 from simpyson.calculator import SimpCalc
+from simpyson.templates import Pulse90
 
 
 def test_custom_pulse_sequence_params():
@@ -67,3 +68,37 @@ def test_standard_params_in_custom_code():
 
     # It should not appear as "variable np"
     assert "variable np" not in par_block
+
+
+# ---------------------------------------------------------------------------
+# Pulse90 — multi-channel padding
+# ---------------------------------------------------------------------------
+
+
+def test_pulse90_single_channel_unpadded():
+    """Single-channel Pulse90 must emit exactly one (rf, phase) pair."""
+    code = Pulse90().generate_code()
+    assert 'pulse $par(pH) $par(plH) $par(phH)\n' in code
+
+
+def test_pulse90_two_channels_padded():
+    """Two-channel Pulse90 must pad the second channel with 0 0."""
+    code = Pulse90(num_channels=2).generate_code()
+    assert 'pulse $par(pH) $par(plH) $par(phH) 0 0\n' in code
+
+
+def test_pulse90_three_channels_padded():
+    code = Pulse90(num_channels=3).generate_code()
+    assert 'pulse $par(pH) $par(plH) $par(phH) 0 0 0 0\n' in code
+
+
+def test_pulse90_multichannel_via_simpcalc():
+    """SimpCalc must pass num_channels to Pulse90 automatically."""
+    calc = SimpCalc(
+        spinsys="channels 1H 13C\nnuclei 1H 13C\nshift 1 5p 0 0 0 0 0\nshift 2 50p 0 0 0 0 0",
+        pulse_sequence='pulse_90',
+        proton_frequency=400e6, spin_rate=10000, sw=20000, np=1024,
+        start_operator='I1x', detect_operator='I2p', method='direct',
+        crystal_file='rep100', gamma_angles=10, verbose=0,
+    )
+    assert 'pulse $par(pH) $par(plH) $par(phH) 0 0' in str(calc)
